pull

+++ hyb/Actievadis/Database/activityFunctions.php
@@ -61,13 +61,12 @@ function addNewActivity($name, $location, $food, $price, $description, $image, $
     $query->bind_param("ssidssss", $nameQ, $locationQ, $foodQ, $priceQ, $descriptionQ, $imageQ, $startTimeQ, $endTimeQ);
     $nameQ = $name;
     $locationQ = $location;
-    $foodQ = $food;
+    $foodQ = ($food) ? true : false;
     $priceQ = $price;
     $descriptionQ = $description;
     $imageQ = $image;
     $startTimeQ = $startTime;
     $endTimeQ = $endTime;
-
     $query->execute();
     $mysqli->close();
 }
@@ -75,12 +74,13 @@ function addNewActivity($name, $location, $food, $price, $description, $image, $
 function deleteActivityById($id)
 {
     $query = "DELETE FROM `activity` WHERE `activity`.`Id` = " . $id;
+    $queryDelSignups = "DELETE FROM signup WHERE activityId = " . $id;
     
     $activity = new activity(getActivityById($id));
     
     if (file_exists("../../Images/".$activity->getImage()))
     {
-        if (db_doQuery($query))
+        if (db_doQuery($query) && db_doQuery($queryDelSignups))
         {
             unlink("../../Images/".$activity->getImage());
         }
